Creating Orderbook

+++ src/models/trade.rs
@@ -3,16 +3,14 @@ use uuid::Uuid;
 
 use crate::schema::{trades};
 
-// TODO implement stop orders
-
 #[derive(Debug, Queryable, Identifiable)]
 pub struct Trade {
     pub id: Uuid,
     pub account_id: Uuid,
     pub order_id: Uuid,
     pub instrument_id: Uuid,
-    pub price: f64,
-    pub exec_qty: i64
+    pub price: f32,
+    pub exec_qty: i32,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
 }
@@ -24,6 +22,6 @@ pub struct NewTrade {
     pub account_id: Uuid,
     pub order_id: Uuid,
     pub instrument_id: Uuid,
-    pub price: f64,
-    pub exec_qty: i64
+    pub price: f32,
+    pub exec_qty: i32
 }
